### 2020-05-21 #2198 @whedon
[PRE REVIEW]: cmstatr: An R Package for Statistical Analysis of Composite Material Data
**Submitting author:** @kloppen (<a href="http://orcid.org/0000-0002-1908-5214">Stefan Kloppenborg</a>)
**Repository:** <a href="https://github.com/ComtekAdvancedStructures/cmstatr" target ="_blank">https://github.com/ComtekAdvancedStructures/cmstatr</a>
**Version:** 0.5.2
