### 2020-05-22 #2198 @kloppen
@usethedata Thanks for volunteering to be the editor for this submission!

I did a cursory search of the reviewers list for people with R as a preferred language and who have listed domains that are somehow related to `cmstatr` and came up with a few names.
- trallard
- myousefi2016
- njtierney
- JonasMoss

This list is certainly not exhaustive. You must have a fairly good idea of which reviewers would be good for reviewing this submission. I'm sure that whichever people you ask will be excellent choices.

I'm looking forward to the review process and receiving some feedback on the software and documentation.
