### 2020-06-01 0a7a426
Updated workflow with changes from r-lib/actions repo
