### 2020-06-01 7f4ec3d
Removed R 3.5 and prior from check workflow
Some part of the toolchain for building vignettes now calls
vignetteInfo(), which was added in R 3.6. As such, this
workflow fails on R 3.5 and previous.
