+++ .github/workflows/R-CMD-check.yaml
@@ -29,9 +29,6 @@ jobs:
         - { os: macOS-latest,   r: 'release',   vdiffr: false }
         - { os: ubuntu-16.04,   r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", vdiffr: true  }
         - { os: ubuntu-16.04,   r: 'oldrel', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", vdiffr: false }
-        - { os: ubuntu-16.04,   r: '3.5', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", vdiffr: false }
-        - { os: ubuntu-16.04,   r: '3.4', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", vdiffr: false }
-        - { os: ubuntu-16.04,   r: '3.3', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", vdiffr: false }
         # r-devel is not available on Ubuntu (ref. r-lib/actions/setup-r/lib/installer)
 
     env:
