### 2020-06-01 5a100b6
Merge R-CMD-check workflow from develop branch
