+++ pkgdown.yml
@@ -1,11 +1,11 @@
 pandoc: 2.7.3
 pkgdown: 1.5.1.9000
-pkgdown_sha: c1db3982037caffe4d487afaef101e49e867cbb4
+pkgdown_sha: aeb1cfb080f566b32da6613b6e12c13eb743daff
 articles:
   adktest: adktest.html
   cmstatr_Graphing: cmstatr_Graphing.html
   cmstatr_Tutorial: cmstatr_Tutorial.html
-last_built: 2020-05-20T12:55Z
+last_built: 2020-06-01T13:36Z
 urls:
   reference: www.cmstatr.net/reference
   article: www.cmstatr.net/articles
