### 2020-06-17 #2265 @JonasMoss
@kloppen Here is the report! Please don't hesitate to ask for details or make comments!

This package is thorough in its documentation and testing and has a clearly laid out reason to exist. I'm impressed by the authors' devotion to showing agreement with CMH-17-1G in almost every function. I deeply appreciate comments such as "The results of this function have been validated against published values in Lawless (1982).", which demonstrates the seriousness of the authors. What's more, it has three vignettes and a unified philosophy. I would have without a doubt used this package if it had been relevant to my work.

Not surprisingly, I only found minor issues with this package.

**Documentation.** The documentation is usually strong in explaining what the arguments are there for. But it's often difficult to understand what different functions *do*. Take `k_equiv`, with the description `k-factors for equivalency testing`. I need an explanation of what k-factors are and what equivalency testing is (which I suspect is used differently in this package than the usual statistical terminology "equivalence test").
To help you understand the problem, compare `k_equiv` with `
nonpara_binomial_rank`, which is much easier to understand and is well-referenced. You state that it finds "distribution-free tolerance limits for large samples" using a named method, and supply a reference for that method. That's what I'd like to see.  However, please fix the reference there.
- [ ] Use [this page](https://www.duo.uio.no/handle/10852/48686) to fix the Guenther reference. In particular, add the permanent link [http://urn.nb.no/URN:NBN:no-52541](http://urn.nb.no/URN:NBN:no-52541) to the package.

I find it difficult to understand, thus verify, the following functions too. (And this problem is compounded when I don't see an example.)
- [ ] `maximum_normed_residual`
- [ ] `equiv_change_mean` needs a reference. Is it a wrapper of a two-sample *t*-test? Then it should say so in Description. What does the `modcv = TRUE` option do? Something described in CMH-17-1G?
- [ ] `equiv_mean_extremum`. Could you explain what this function does, preferably with a reference?
- [ ]  `k_factor_normal`: You should define what the k factor is and provide a reference or maybe remove it from the public API.
- [ ] `basis_values`: Needs a reference to tolerance interval and its definition.
- [ ] Add reference to tolerance intervals, where needed, to explain what you are doing.

There might be other functions too with too little of "what does the function do" too. Please be aware that I'm not asking for you to rewrite the documentation (which is good), I just need some more information in some circumstances. In particular, I have to be able to verify if that the exported functions do what they say they do, this is one of the unchecked checkboxes -- and then I have to understand what they are supposed to do. ;)
- [ ] Explain what the functions do.
- Spelling and writing. Some minor issues here too. I haven't hunted for spelling errors, and I don't think they are important.
	- [ ] Anderson-Darling -> Anderson--Darling (does Rmarkdown support en-dashes though?)
	- [ ] Hanson-Koopmans -> Hanson--Koopmans
	- [ ] Please write out "CV" at least once when you use it in the documenation; this is true for any abbreviation. (E.g. `
calc_cv_star`; which should cross-reference `cv` too.)
    - [ ] Output of `equiv_change_mean`: Qualificaiton -> Qualification
    - [ ] Some functions have very minor spacing issues in the ouput, for instance `levene_test`.
    - [ ] 	`basis`: preform -> perform.
- [ ] Terminology.
	- [ ] In, say, the Anderson--Darling tests, you write "significance" when you report a *p*-value.
	- [ ] Likewise, you write in the documentation that "The significance level is calculated assuming that the parameters of the distribution are unknown; these parameters are estimate from the data." The proper terminology is "*p*-value"; not "significance level", which refers to a significance cutoff (i.e. `alpha`.)
- [ ] Verification:
	- [ ] Your normal Anderson--Darling test does not agree with `nortest::ad.test`. Why is that so?
- [ ] Examples. It would be nice to have examples for all exported functions; this should be easy; just copy from your `test_that` folder. In particular, the following functions lack examples.
	- [ ] `levene_test`
	- [ ] `maximum_normed_residual`
	- [ ] `nonpara_binomial_rank`

Calling
> devtools::install_github("ComtekAdvancedStructures/cmstatr", build_vignettes = TRUE, build_opts = c("--no-resave-data", "--no-manual"))

fails when `rmarkdown`  `dplyr`, and `tidyr` aren't installed. I suggest adding `install.packages("rmarkdown")`, and the same for the rest, to the installation instructions or, (ii) add rmarkdown to Imports instead of Suggests. (iii) It is possible to add checks in the vignette code to find out if a package has been installed, then ask to the user to install it right there and then. Don't do it automatically though -- CRAN won't accept it! In my opinion, this third option is not worth the trouble.

- [ ] Installation. Fix the `rmarkdown`, `dplyr`, `tidyr` issue.

> Design Values are determined such that, with 95% confidence, they are either the 99% or 90% one-sided lower confidence bound of the material strength, depending on the type of structure.

I was strongly confused by this passage, and I think it should be reworded or removed. An option is to use only the language of tolerance intervals/bounds, and provide a reference and introduction to their definition and interpretation. Tolerance intervals are too convoluted to properly define in this paper (or the documentation), but they should be referenced by their proper name so that people (that is, people like me! :p) can easily check if they are implemented correctly. Moreover, I suggest to use the terminology "content" and "confidence level" for `p` and `conf`. Since you only care about lower confidence levels, I suggest you remove references to confidence intervals and replace them with confidence levels instead; it is confusing that several functions take `conf` as an argument but fails to return a confidence interval.

It is nice that you repeatedly state in the documentation that p = 0.9 and conf = 0.95 corresponds to a B-Basis and so on, you should also state what they are supposed to calculate when p != 0.9 and conf != 0.95. (That is, the lower limits of a one-sided tolerance interval with content p and confidence level 0.95.)

A reference for the to me surprising calculations used in e.g. `basis_normal` and `k_factor_normal` is Statistical Tolerance Regions: Theory, Applications, and Computation (Krishnamoorthy, Mathew, 2009), Section 1.2. But you probably know better references than I do!

- [ ] Documenation: Reference to tolerance bounds.

The rest of the software paper looks fine to me!
