### 2020-06-17 #2265 @kloppen
@JonasMoss Thank you very much for your comments. You clearly spent considerable effort in conducting a thorough review of the paper and the package. I really appreciate it!

I'll start addressing your comments and will let you know when I'm done. You were quite clear in what you thought should be changed, so I don't anticipate that I'll need any clarification of your comments, but if something comes up, I'll be sure to ask.

Thanks again!
