### 2020-06-17 7dc781b
Updated installation instructions
