### 2020-06-17 fb6a09d
Added links to online version of vignettes
