+++ README.Rmd
@@ -51,7 +51,8 @@ carbon.fabric.2 %>%
 
 For more examples of usage of the `cmstatr` package,
 see the tutorial vignette, which can be
-loaded as follows, once the package is installed:
+[viewed online](https://www.cmstatr.net/articles/cmstatr_Tutorial.html),
+or can be loaded as follows, once the package is installed:
 
 ```r
 vignette("cmstatr_Tutorial")
@@ -59,7 +60,9 @@ vignette("cmstatr_Tutorial")
 
 There is also a vignette showing some examples of the types of graphs
 that are typically produced when analyzing composite materials.
-You can load this vignette with:
+You can view this
+[vignette online](https://www.cmstatr.net/articles/cmstatr_Graphing.html),
+or you can load this vignette with:
 
 ```r
 vignette("cmstatr_Graphing")
