### 2020-06-17 c36c520
Updated confusing sentence; added reference to bibliography
