+++ paper.bib
@@ -19,7 +19,6 @@
   year = {1970},
   DOI = {10.2307/2682297}
 }
-
 @article{Scholz_Stephens_1987,
   title={K-Sample Anderson-Darling Tests},
   volume={82},
@@ -99,3 +98,12 @@
     pages = {1686},
     doi = {10.21105/joss.01686},
   }
+@book{Krishnamoorthy_Mathew_2008,
+  place={Hoboken},
+  title={Statistical Tolerance Regions: Theory, Applications, and Computation},
+  ISBN={9780470473900},
+  doi={10.1002/9780470473900},
+  publisher={John Wiley & Sons},
+  author={Krishnamoorthy, K. and Mathew, Thomas},
+  year={2008}
+}
