### 2020-06-17 eac7522
Added reference to Krishnamoorthy in documentation; fixed false-positive spelling errors
