### 2020-06-18 92ffa21
Improved MNR documentation
