+++ R/mnr.R
@@ -1,5 +1,12 @@
 #' Detect outliers using the maximum normed residual method
 #'
+#' @description
+#' This function detects outliers using the maximum normed residual
+#' method described in CMH-17-1G. This method identifies a value
+#' as an outlier if the absolute difference between the value and
+#' the sample mean divided by the sample standard deviation
+#' exceeds a critical value.
+#'
 #' @param data a data.frame
 #' @param x the variable in the data.frame for which to find the MNR
 #'          or a vector if \code{data=NULL}
@@ -15,6 +22,18 @@
 #' The maximum normed residual test is a test for outliers. The test statistic
 #' is given in CMH-17-1G. Outliers are identified in the returned object.
 #'
+#' The maximum normed residual test statistic is defined as:
+#'
+#' \deqn{MNR = max \frac{\left| xi - \bar{x} \right|}{s} }{
+#'   MNR = max | x_i- x_bar | / s }
+#'
+#' When the value of the MNR test statistic exceeds the critical value
+#' defined in Section 8.3.3.1 of CMH-17-1G, the corresponding value
+#' is identified as an outlier. It is then removed from the sample, and
+#' the test statistic is computed again and compared with the critical
+#' value corresponding with the new sample. This process is repeated until
+#' no values are identified as outliers.
+#'
 #' @return an object of class \code{mnr}
 #' This object has the following fields:
 #'  \item{\code{call}}{the expression used to call this function}
@@ -25,6 +44,38 @@
 #'  \item{\code{outliers}}{a data.frame containing the \code{index} and
 #'                         \code{value} of each of the identified outliers}
 #'  \item{\code{n_outliers}}{the number of outliers found}
+#'
+#' @examples
+#' library(dplyr)
+#'
+#' carbon.fabric.2 %>%
+#'   filter(test=="FC" & condition=="ETW2" & batch=="A") %>%
+#'   maximum_normed_residual(strength)
+#'
+#' ## Call:
+#' ## maximum_normed_residual(data = ., x = strength)
+#' ##
+#' ## MNR =  1.958797  ( critical value =  1.887145 )
+#' ##
+#' ## Outliers:
+#' ##   Index  Value
+#' ##       6  44.26
+#'
+#' carbon.fabric.2 %>%
+#'   filter(test=="FC" & condition=="ETW2" & batch=="B") %>%
+#'   maximum_normed_residual(strength)
+#'
+#' ## Call:
+#' ## maximum_normed_residual(data = ., x = strength)
+#' ##
+#' ## MNR =  1.469517  ( critical value =  1.887145 )
+#' ##
+#' ## No outliers detected
+#'
+#' @references
+#' “Composite Materials Handbook, Volume 1. Polymer Matrix Composites
+#' Guideline for Characterization of Structural Materials,” SAE International,
+#' CMH-17-1G, Mar. 2012.
 #'
 #' @importFrom rlang eval_tidy enquo
 #'
