+++ man/maximum_normed_residual.Rd
@@ -27,7 +27,11 @@ This object has the following fields:
  \item{\code{n_outliers}}{the number of outliers found}
 }
 \description{
-Detect outliers using the maximum normed residual method
+This function detects outliers using the maximum normed residual
+method described in CMH-17-1G. This method identifies a value
+as an outlier if the absolute difference between the value and
+the sample mean divided by the sample standard deviation
+exceeds a critical value.
 }
 \details{
 \code{data} is an optional argument. If \code{data} is given, it
@@ -38,4 +42,49 @@ value of \code{x} is expected to be a variable within \code{data}. If
 
 The maximum normed residual test is a test for outliers. The test statistic
 is given in CMH-17-1G. Outliers are identified in the returned object.
+
+The maximum normed residual test statistic is defined as:
+
+\deqn{MNR = max \frac{\left| xi - \bar{x} \right|}{s} }{
+  MNR = max | x_i- x_bar | / s }
+
+When the value of the MNR test statistic exceeds the critical value
+defined in Section 8.3.3.1 of CMH-17-1G, the corresponding value
+is identified as an outlier. It is then removed from the sample, and
+the test statistic is computed again and compared with the critical
+value corresponding with the new sample. This process is repeated until
+no values are identified as outliers.
+}
+\examples{
+library(dplyr)
+
+carbon.fabric.2 \%>\%
+  filter(test=="FC" & condition=="ETW2" & batch=="A") \%>\%
+  maximum_normed_residual(strength)
+
+## Call:
+## maximum_normed_residual(data = ., x = strength)
+##
+## MNR =  1.958797  ( critical value =  1.887145 )
+##
+## Outliers:
+##   Index  Value
+##       6  44.26
+
+carbon.fabric.2 \%>\%
+  filter(test=="FC" & condition=="ETW2" & batch=="B") \%>\%
+  maximum_normed_residual(strength)
+
+## Call:
+## maximum_normed_residual(data = ., x = strength)
+##
+## MNR =  1.469517  ( critical value =  1.887145 )
+##
+## No outliers detected
+
+}
+\references{
+“Composite Materials Handbook, Volume 1. Polymer Matrix Composites
+Guideline for Characterization of Structural Materials,” SAE International,
+CMH-17-1G, Mar. 2012.
 }
