### 2020-06-18 685aeb9
Updated Anderson-Darling to Anderson--Darling
