+++ paper.bib
@@ -20,7 +20,7 @@
   DOI = {10.2307/2682297}
 }
 @article{Scholz_Stephens_1987,
-  title={K-Sample Anderson-Darling Tests},
+  title={K-Sample Anderson--Darling Tests},
   volume={82},
   number={399},
   journal={Journal of the American Statistical Association},
