### 2020-06-19 0b26e1d
Updated documentation for k_equiv and equiv_change_mean
