### 2020-06-19 7fdf604
Updated equiv_change_mean documentation; minor tweak to docs for equiv_mean_extremum
