+++ man/equiv_change_mean.Rd
@@ -94,6 +94,7 @@ should be used. Defaults to \code{FALSE}}
 \description{
 Checks for change in the mean value between a qualification data set and
 a sample. This is normally used to check for properties such as modulus.
+This function is a wrapper for a two-sided t--test.
 }
 \details{
 There are several optional arguments to this function. Either (but not both)
@@ -109,11 +110,16 @@ This function uses a two-sided t-test to determine if there is a difference
 in the mean value of the qualification data and the sample. A pooled
 standard deviation is used in the t-test. The procedure is per CMH-17-1G.
 
-If the option \code{modcv = TRUE} is set, standard deviation of the
-qualification data is replaced with CV* times \code{mean_qual} (which may
-be passed as an argument or internally calculated from \code{data_qual}.
-
-When \code{modcv = TRUE}, CV* is calculated as follows:
+If \code{modcv} is TRUE, the standard deviation used to calculate the
+thresholds will be replaced with a standard deviation calculated
+using the Modified Coefficient of Variation (CV) approach.
+The Modified CV approach is a way of adding extra variance to the
+qualification data in the case that the qualification data has less
+variance than expected, which sometimes occurs when qualification testing
+is performed in a short period of time.
+Using the Modified CV approach, the standard deviation is calculated by
+multiplying \code{CV_star * mean_qual} where \code{mean_qual} is either the
+value supplied or the value calculated by \code{mean(data_qual)} and
 \eqn{CV* = 0.06} if \eqn{CV < 0.04}, \eqn{CV* = cv / 2 + 0.04}
 if \eqn{0.04 <= cv <= 0.08} and \eqn{CV* = CV} otherwise.
 
@@ -139,4 +145,13 @@ equiv_change_mean(alpha = 0.05, n_sample = 9, mean_sample = 9.02,
 ##           Result               PASS
 ##    Passing Range       8.856695 to 9.623305
 
+}
+\references{
+“Composite Materials Handbook, Volume 1. Polymer Matrix Composites
+Guideline for Characterization of Structural Materials,” SAE International,
+CMH-17-1G, Mar. 2012.
+}
+\seealso{
+\code{\link{calc_cv_star}}
+\code{\link[stats]{t.test}}
 }
