### 2020-06-19 ab6b5d0
Em dash for Hanson--Koopmans
