+++ R/basis.R
@@ -59,7 +59,7 @@ k_factor_normal <- function(n, p = 0.90, conf = 0.95) {
 #'                 if desired.
 #' @param modcv a logical value indicating whether the modified CV approach
 #'              should be used. Only applicable to pooling methods.
-#' @param method the method for Hanson-Koopmans nonparametric basis values.
+#' @param method the method for Hanson--Koopmans nonparametric basis values.
 #'               should be "optimum-order" for B-Basis and "woodward-frawley"
 #'               for A-Basis.
 #'
@@ -122,7 +122,7 @@ k_factor_normal <- function(n, p = 0.90, conf = 0.95) {
 #' for these diagnostic tests.
 #'
 #' \code{basis_hk_ext} calculates the basis value using the Extended
-#' Hanson-Koopmans method, as described in CMH-17-1G and Vangel (1994).
+#' Hanson--Koopmans method, as described in CMH-17-1G and Vangel (1994).
 #' For nonparametric distributions, this function should be used for samples
 #' up to n=28 for B-Basis and up to \eqn{n=299} for A-Basis.
 #' This method uses a pair of order statistics to determine the basis value.
@@ -877,10 +877,10 @@ hk_ext_h <- function(z, n, i, j, p) {
   qb + int$value
 }
 
-#' Calculate values related to the Extended Hanson-Koopmans method
+#' Calculate values related to the Extended Hanson---Koopmans method
 #'
 #' @description
-#' Calculates values related to the Extended Hanson-Koopmans method
+#' Calculates values related to the Extended Hanson--Koopmans method
 #' as described by Vangel (1994).
 #'
 #' @param n the sample size
@@ -902,7 +902,7 @@ hk_ext_h <- function(z, n, i, j, p) {
 #' tolerance limits based on consecutive order statistics.
 #' Vangel (1994) extends this method using non-consecutive order statistics.
 #'
-#' The extended Hanson-Koopmans method calculates a tolerance limit
+#' The extended Hanson--Koopmans method calculates a tolerance limit
 #' (basis value) based on two order statistics and a weighting value
 #' \code{z}. The value of \code{z} is based on the sample size, which
 #' order statistics are selected, the desired quantile and the desired
