### 2020-06-19 ffd31d5
Updated documentation to spell out coefficient of variation
