+++ R/levene.R
@@ -37,6 +37,10 @@
 #' Where \eqn{mi} is median of the \eqn{ith} group. An F-Test is then
 #' performed on the transformed data.
 #'
+#' When \code{modcv=TRUE}, the data from each group is first transformed
+#' according to the modified coefficient of variation (CV) rules before
+#' performing Levene's test.
+#'
 #' @references
 #' “Composite Materials Handbook, Volume 1. Polymer Matrix Composites
 #' Guideline for Characterization of Structural Materials,” SAE International,
@@ -44,6 +48,10 @@
 #'
 #' @importFrom rlang enquo eval_tidy
 #' @importFrom stats var.test median pf
+#'
+#' @seealso
+#' \code{\link{calc_cv_star}}
+#' \code{\link{transform_mod_cv}}
 #'
 #' @export
 levene_test <- function(data = NULL, x, groups, alpha = 0.05, modcv = FALSE) {
