+++ man/calc_cv_star.Rd
@@ -13,7 +13,8 @@ calc_cv_star(cv)
 The value of the modified CV
 }
 \description{
-This function calculates the Modified CV based on a (unmodified) CV.
+This function calculates the Modified coefficient of variation (CV)
+based on a (unmodified) CV.
 The modified CV is calculated based on the rules in CMH-17-1G. Those
 rules are:
 
@@ -27,4 +28,7 @@ rules are:
 "Composite Materials Handbook, Volume 1. Polymer Matrix Composites
 Guideline for Characterization of Structural Materials,"
 SAE International, CMH-17-1G, Mar. 2012.
+}
+\seealso{
+\code{\link{cv}}
 }
