### 2020-06-21 0b115ab
Improved documentation of k_factor_normal and basis_values
