### 2020-06-22 668ee9b
Fixed spacing issues; standardized (mostly) the way that printing is done
