### 2020-06-22 78c1f82
Fixed minor formatting errors in documentation
