+++ R/basis.R
@@ -54,6 +54,25 @@
 #' @seealso
 #' \code{\link{basis_normal}}
 #'
+#' @examples
+#' kb <- k_factor_normal(n = 10, p = 0.9, conf = 0.95)
+#' print(kb)
+#'
+#' ## [1] 2.35464
+#'
+#' # This can be used to caclulate the B-Basis if
+#' # the sample mean and sample standard deviation
+#' # is known, and data is assumed to be normally
+#' # distributed
+#'
+#' sample_mean <- 90
+#' sample_sd <- 5.2
+#' print("B-Basis:")
+#' print(sample_mean - sample_sd * kb)
+#'
+#' ## [1] B-Basis:
+#' ## [1] 77.75587
+#'
 #' @importFrom stats qnorm qt
 #'
 #' @export
@@ -307,6 +326,54 @@ k_factor_normal <- function(n, p = 0.90, conf = 0.95) {
 #' W. Meeker, G. Hahn, and L. Escobar, Statistical Intervals: A Guide
 #' for Practitioners and Researchers, Second Edition.
 #' Hoboken: John Wiley & Sons, 2017.
+#'
+#' @examples
+#' library(dplyr)
+#'
+#' # A single-point basis value can be calculated as follows
+#' # in this example, three failed diagnostic tests are
+#' # overriden.
+#'
+#' carbon.fabric %>%
+#'   filter(test == "FC") %>%
+#'   filter(condition == "RTD") %>%
+#'   basis_normal(strength, batch,
+#'                override = c("outliers",
+#'                             "outliers_within_batch",
+#'                             "anderson_darling_normal"))
+#'
+#' ## Call:
+#' ## basis_normal(data = ., x = strength, batch = batch,
+#' ##     override = c("outliers", "outliers_within_batch",
+#' ##    "anderson_darling_normal"))
+#' ##
+#' ## Distribution:  Normal 	( n = 18 )
+#' ## The following diagnostic tests were overridden:
+#' ##     `outliers`,
+#' ##     `outliers_within_batch`,
+#' ##     `anderson_darling_normal`
+#' ## B-Basis:   ( p = 0.9 , conf = 0.95 )
+#' ## 76.94656
+#'
+#' # A set of pooled basis values can also be calculated
+#' # using the pooled standard deviation method, as follows.
+#' # In this example, one failed diagnostic test is overriden.
+#' carbon.fabric %>%
+#'   filter(test == "WT") %>%
+#'   basis_pooled_sd(strength, condition, batch,
+#'                   override = c("outliers_within_batch"))
+#'
+#' ## Call:
+#' ## basis_pooled_sd(data = ., x = strength, groups = condition,
+#' ##                 batch = batch, override = c("outliers_within_batch"))
+#' ##
+#' ## Distribution:  Normal - Pooled Standard Deviation 	( n = 54, r = 3 )
+#' ## The following diagnostic tests were overridden:
+#' ##     `outliers_within_batch`
+#' ## B-Basis:   ( p = 0.9 , conf = 0.95 )
+#' ## CTD  127.6914
+#' ## ETW  125.0698
+#' ## RTD  132.1457
 #'
 #' @name basis
 NULL
@@ -923,7 +990,7 @@ hk_ext_h <- function(z, n, i, j, p) {
   qb + int$value
 }
 
-#' Calculate values related to the Extended Hanson---Koopmans method
+#' Calculate values related to the Extended Hanson--Koopmans method
 #'
 #' @description
 #' Calculates values related to the Extended Hanson--Koopmans method
@@ -955,7 +1022,13 @@ hk_ext_h <- function(z, n, i, j, p) {
[…]
