### 2020-06-24 bc6d412
Added discussion of AD test difference with nortest and cleaned up terminology
