+++ R/adtest.R
@@ -3,12 +3,13 @@
 #'
 #' @description
 #' Calculates the Anderson--Darling test statistic for a sample given
-#' a particular distribution, and computes the significance level.
+#' a particular distribution, and determines whether to reject the
+#' hypothesis that a sample is drawn from that distribution.
 #'
 #' @param data a data.frame-like object (optional)
 #' @param x a numeric vector or a variable in the data.frame
-#' @param alpha the required significance to conclude that the data follows
-#'   the specified distribution. Defaults to 0.05.
+#' @param alpha the required significance level of the test.
+#'              Defaults to 0.05.
 #'
 #' @return
 #' an object of class \code{anderson_darling}. This object has the following
@@ -19,10 +20,10 @@
 #' \item{\code{data}}{a copy of the data analyzed}
 #' \item{\code{n}}{the number of observations in the sample}
 #' \item{\code{A}}{the Anderson--Darling test statistic}
-#' \item{\code{osl}}{the significance level, assuming the
+#' \item{\code{osl}}{the observed significance level (p-value),
+#'   assuming the
 #'   parameters of the distribution are estimated from the data}
-#' \item{\code{alpha}}{the required significance level for the test to
-#'  conclude that the data is drawn from the specified distribution.
+#' \item{\code{alpha}}{the required significance level for the test.
 #'  This value is given by the user.}
 #' \item{\code{reject_distribution}}{a logical value indicating whether
 #'  the hypothesis that the data is drawn from the specified distribution
@@ -32,7 +33,8 @@
 #' The Anderson--Darling test statistic is calculated for the distribution
 #' given by the user.
 #'
-#' The significance level is calculated assuming that the parameters
+#' The observed significance level (OSL), or p-value, is calculated assuming
+#' that the parameters
 #' of the distribution are unknown; these parameters are estimate from the
 #' data.
 #'
@@ -48,12 +50,30 @@
 #' test statistic given a Weibull distribution with shape and scale parameters
 #' estimate from the data using a maximum likelihood estimate.
 #'
+#' The test statistic, \code{A}, is modified to account for
+#' the fact that the parameters of the population are not known,
+#' but are instead estimated from the sample. This modification is
+#' a function of the sample size only, and is different for each
+#' distribution (normal/lognormal or Weibull). Several such modifications
+#' have been proposed. This function uses the modification published in
+#' Stephens (1974), Lawless (1982) and CMH-17-1G. Some other implementations
+#' of the Anderson-Darling test, such as the implementation in the
+#' \code{nortest} package, use other modifications, such as the one
+#' published in D'Agostino and Stephens (1986). As such, the p-value
+#' reported by this function may differ from the p-value reported
+#' by implementations of the Anderson--Darling test that use
+#' different modifiers. Only the unmodified
+#' test statistic is reported in the result of this function, but
+#' the modified test statistic is used to compute the OSL (p-value)
+#' and when rejecting or failing to reject the null hypothesis.
+#'
 #' This function uses the formula for observed significance
-#' level published in CMH-17-1G. These formulae depend on the particular
+#' level (OSL) published in CMH-17-1G. These formulae depend on the particular
 #' distribution used.
 #'
 #' The results of this function have been validated against
 #' published values in Lawless (1982).
+#'
 #'
 #' @references
[…]
