+++ man/glance.anderson_darling.Rd
@@ -19,10 +19,10 @@ columns:
 \item{\code{dist}}{the distribution used}
 \item{\code{n}}{the number of observations in the sample}
 \item{\code{A}}{the Anderson--Darling test statistic}
-\item{\code{osl}}{the significance level, assuming the
+\item{\code{osl}}{the observed significance level (p-value),
+    assuming the
     parameters of the distribution are estimated from the data}
- \item{\code{alpha}}{the required significance level for the test to
-   conclude that the data is drawn from the specified distribution.
+ \item{\code{alpha}}{the required significance level for the test.
    This value is given by the user.}
 \item{\code{reject_distribution}}{a logical value indicating whether
    the hypothesis that the data is drawn from the specified distribution
