+++ man/anderson_darling.Rd
@@ -18,8 +18,8 @@ anderson_darling_weibull(data = NULL, x, alpha = 0.05)
 
 \item{x}{a numeric vector or a variable in the data.frame}
 
-\item{alpha}{the required significance to conclude that the data follows
-the specified distribution. Defaults to 0.05.}
+\item{alpha}{the required significance level of the test.
+Defaults to 0.05.}
 }
 \value{
 an object of class \code{anderson_darling}. This object has the following
@@ -30,10 +30,10 @@ fields.
 \item{\code{data}}{a copy of the data analyzed}
 \item{\code{n}}{the number of observations in the sample}
 \item{\code{A}}{the Anderson--Darling test statistic}
-\item{\code{osl}}{the significance level, assuming the
+\item{\code{osl}}{the observed significance level (p-value),
+  assuming the
   parameters of the distribution are estimated from the data}
-\item{\code{alpha}}{the required significance level for the test to
- conclude that the data is drawn from the specified distribution.
+\item{\code{alpha}}{the required significance level for the test.
  This value is given by the user.}
 \item{\code{reject_distribution}}{a logical value indicating whether
  the hypothesis that the data is drawn from the specified distribution
@@ -41,13 +41,15 @@ fields.
 }
 \description{
 Calculates the Anderson--Darling test statistic for a sample given
-a particular distribution, and computes the significance level.
+a particular distribution, and determines whether to reject the
+hypothesis that a sample is drawn from that distribution.
 }
 \details{
 The Anderson--Darling test statistic is calculated for the distribution
 given by the user.
 
-The significance level is calculated assuming that the parameters
+The observed significance level (OSL), or p-value, is calculated assuming
+that the parameters
 of the distribution are unknown; these parameters are estimate from the
 data.
 
@@ -63,8 +65,25 @@ The function \code{anderson_darling_weibull} computes the Anderson--Darling
 test statistic given a Weibull distribution with shape and scale parameters
 estimate from the data using a maximum likelihood estimate.
 
+The test statistic, \code{A}, is modified to account for
+the fact that the parameters of the population are not known,
+but are instead estimated from the sample. This modification is
+a function of the sample size only, and is different for each
+distribution (normal/lognormal or Weibull). Several such modifications
+have been proposed. This function uses the modification published in
+Stephens (1974), Lawless (1982) and CMH-17-1G. Some other implementations
+of the Anderson-Darling test, such as the implementation in the
+\code{nortest} package, use other modifications, such as the one
+published in D'Agostino and Stephens (1986). As such, the p-value
+reported by this function may differ from the p-value reported
+by implementations of the Anderson--Darling test that use
+different modifiers. Only the unmodified
+test statistic is reported in the result of this function, but
+the modified test statistic is used to compute the OSL (p-value)
+and when rejecting or failing to reject the null hypothesis.
+
 This function uses the formula for observed significance
-level published in CMH-17-1G. These formulae depend on the particular
+level (OSL) published in CMH-17-1G. These formulae depend on the particular
 distribution used.
 
 The results of this function have been validated against
@@ -82,8 +101,8 @@ carbon.fabric \%>\%
 ##
 ## Distribution:  Normal ( n = 18 )
 ## Test statistic:  A = 0.9224776
-## Significance:  0.01212193  (assuming unknown parameters)
-## Conclusion: Sample is not drawn from a Normal distribution (alpha = 0.05 )
+## OSL (p-value):  0.01212193  (assuming unknown parameters)
[…]

[…]
