### 2020-06-24 6e88058
General documentation cleanup
