+++ paper.bib
@@ -60,7 +60,7 @@
 }
 @report{STAT-17,
   place={Horsham, PA},
-  title={CMH-17 Statistical Analysis for B-Basis and A-Basis Values},
+  title={CMH-17 Statistical Analysis for {B}-Basis and {A}-Basis Values},
   number={STAT-17 Rev 5},
   institution={Materials Sciences Corporation},
   author={{Materials Sciences Corporation}},
@@ -75,7 +75,7 @@
   year={2008}
 }
 @report{CMH17-STATS,
-  title={CMH17 Statistical Analysis Program for B-Basis & A-Basis Values},
+  title={CMH17 Statistical Analysis Program for {B}-Basis & {A}-Basis Values},
   number={CMH17-STATS v2011 1.1},
   institution={Witchita State University},
   author={Suresh Keshavanarayana},
@@ -106,4 +106,13 @@
   publisher={John Wiley & Sons},
   author={Krishnamoorthy, K. and Mathew, Thomas},
   year={2008}
+}
+@book{Meeker_Hahn_Escobar_2017,
+  place={Hoboken},
+  edition={Second Edition},
+  title={Statistical Intervals: A Guide for Practitioners and Researchers},
+  ISBN={978-0-471-68717-7},
+  publisher={John Wiley & Sons},
+  author={Meeker, William and Hahn, Gerald and Escobar, Luis},
+  year={2017}
 }
