### 2020-06-24 c46e5e6
Added alpha to mnr object; updated print and glance methods
