### 2020-06-25 cf8d83d
Incorporated Billy Cheng's comments on the documentation
