### 2020-06-25 #2265 @kloppen
Thanks again for your review and comments, @JonasMoss

I've addressed your comments. I've created a separate branch in the `cmstatr` repo for this, called [`joss-review`](https://github.com/ComtekAdvancedStructures/cmstatr/tree/feature/joss-review). There is a brief response to each of your comments below, and in most cases a commit hash corresponding to the actual change that was made to address the comment. In some cases, there were several commits that affected the same part of the documentation (eg. fixing a spelling error in one commit, then restructuring a sentence in another), so I'm not sure if it's easier to look at the individual commits, or to just compare the branch in which the changes have been made (the [`joss-review`](https://github.com/ComtekAdvancedStructures/cmstatr/tree/feature/joss-review) branch) with `master` to see everything that's changed. You should be able to compare those two branches here:

[https://github.com/ComtekAdvancedStructures/cmstatr/compare/feature/joss-review](https://github.com/ComtekAdvancedStructures/cmstatr/compare/feature/joss-review)

If you're satisfied with the changes and the responses below, I'll merge the [`joss-review`](https://github.com/ComtekAdvancedStructures/cmstatr/tree/feature/joss-review) branch into `master` and increment the version number of the package to 0.6.0 (and ask whedon to re-build the paper). If you'd like any additional changes, or would like to discuss any other aspects of the paper or the package, please let me know.

> This package is thorough in its documentation and testing and has a clearly
> laid out reason to exist. I'm impressed by the authors' devotion to showing
> agreement with CMH-17-1G in almost every function. I deeply appreciate
> comments such as "The results of this function have been validated against
> published values in Lawless (1982).", which demonstrates the seriousness of
> the authors. What's more, it has three vignettes and a unified philosophy.
> I would have without a doubt used this package if it had been relevant to my work.

> Not surprisingly, I only found minor issues with this package.

# Documentation
> The documentation is usually strong in explaining what the arguments are there for.
> But it's often difficult to understand what different functions do. Take k_equiv,
> with the description k-factors for equivalency testing. I need an explanation of
> what k-factors are and what equivalency testing is (which I suspect is used
> differently in this package than the usual statistical terminology "equivalence test").
> To help you understand the problem, compare k_equiv with nonpara_binomial_rank,
> which is much easier to understand and is well-referenced. You state that it finds
> "distribution-free tolerance limits for large samples" using a named method, and
> supply a reference for that method. That's what I'd like to see. However, please
> fix the reference there.
>
> - Use this page to fix the Guenther reference. In particular, add the permanent
>   link http://urn.nb.no/URN:NBN:no-52541 to the package.

Commit [0b26e1d](https://github.com/ComtekAdvancedStructures/cmstatr/commit/0b26e1df67988cf492af2f6c8ea10700c9d9c881): I've updated the documentation for both `k_equiv` and `equiv_mean_extremum`.

Commit [4042e15](https://github.com/ComtekAdvancedStructures/cmstatr/commit/4042e1557b6e2bbeffb527647156243e6498c397): Added URL to documentation for `nonpara_binomial_rank` function. I've left the reference in paper.bib pointing to the JSTOR URL and the American Statistician journal so that I can keep the DOI, which does refer to the published version of this paper, not the preprint.

>
> I find it difficult to understand, thus verify, the following functions too.
> (And this problem is compounded when I don't see an example.)
>
> - maximum_normed_residual

Commit [92ffa21](https://github.com/ComtekAdvancedStructures/cmstatr/commit/92ffa218c90976e9a706d6a0a4f750ad0a3a6376)

>
> - equiv_change_mean needs a reference. Is it a wrapper of a two-sample t-test?
>   Then it should say so in Description. What does the modcv = TRUE option do?
>   Something described in CMH-17-1G?

Commit [7fdf604](https://github.com/ComtekAdvancedStructures/cmstatr/commit/7fdf60428eb33256764029e2459d51d7a87e7844)

>
> - equiv_mean_extremum. Could you explain what this function does,
>   preferably with a reference?

Commit [0b26e1d](https://github.com/ComtekAdvancedStructures/cmstatr/commit/0b26e1df67988cf492af2f6c8ea10700c9d9c881)

>
> - k_factor_normal: You should define what the k factor is and provide a
>   reference or maybe remove it from the public API.
> - basis_values: Needs a reference to tolerance interval and its definition.
>

Both addressed in commit [0b115ab](https://github.com/ComtekAdvancedStructures/cmstatr/commit/0b115ab261e253aa7b7f28340f444fcec57ab470)

>
> - Add reference to tolerance intervals, where needed, to explain what you are doing.
>

These have been added throughout the documentation, as needed.

> There might be other functions too with too little of "what does the function do"
> too. Please be aware that I'm not asking for you to rewrite the documentation
> (which is good), I just need some more information in some circumstances.
> In particular, I have to be able to verify if that the exported functions do
> what they say they do, this is one of the unchecked checkboxes -- and then I have
> to understand what they are supposed to do. ;)
>
> - Explain what the functions do.
> - Spelling and writing. Some minor issues here too. I haven't hunted for
>   spelling errors, and I don't think they are important.
>   - Anderson-Darling -> Anderson--Darling (does Rmarkdown support en-dashes though?)

Commit [685aeb9](https://github.com/ComtekAdvancedStructures/cmstatr/commit/685aeb9e636a02e886c2724faf6476967495c543)

>
>   - Hanson-Koopmans -> Hanson--Koopmans
>

Commit [ab6b5d0](https://github.com/ComtekAdvancedStructures/cmstatr/commit/ab6b5d01c8ac5ce6828c6ac3fed02b308c7d2c05)

>
>   - Please write out "CV" at least once when you use it in the documenation;
>     this is true for any abbreviation. (E.g. calc_cv_star; which should cross-reference cv too.)

Commit [ffd31d5](https://github.com/ComtekAdvancedStructures/cmstatr/commit/ffd31d58b69b1074d8e91a2ecc68c048cb24cdb4)

>
>   - Output of equiv_change_mean: Qualificaiton -> Qualification
>

Commit [cebef8e](https://github.com/ComtekAdvancedStructures/cmstatr/commit/cebef8ed6f80e6e23275ff222bd6601f555dc8c5)

>
>   - Some functions have very minor spacing issues in the ouput, for instance levene_test.
>

Commit [668ee9b](https://github.com/ComtekAdvancedStructures/cmstatr/commit/668ee9be38e1a2afb7ef24c82bf5256725905e7d)

>
>   - basis: preform -> perform.
>

Commit [b859093](https://github.com/ComtekAdvancedStructures/cmstatr/commit/b8590939b08d8149996fa584231f6fc4fbdd9618)

>
> - Terminology.
>   - In, say, the Anderson--Darling tests, you write "significance" when you report a p-value.
>   - Likewise, you write in the documentation that "The significance level is calculated assuming
>     that the parameters of the distribution are unknown; these parameters are estimate from the
>     data." The proper terminology is "p-value"; not "significance level", which refers to a
>     significance cutoff (i.e. alpha.)
>

Commit [bc6d412](https://github.com/ComtekAdvancedStructures/cmstatr/commit/bc6d4126b3ecdcf7541db9cdf88aa9e18dfe8e1f): In the case of the Anderson--Darling tests, I've tried to clear up the terminology, primarily using the term observed significance level, which is the term used by the CMH-17-1G handbook, and also the existing software in use. My understanding is that observed significance level (OSL) is a much less common term that means the same thing as p-value. I agree that calling this p-value, rather than the much less common name observed significance level, would be the more precise terminology, but I do think that there is value in using the terminology that the likely users of this package are accustomed to.

>
> - Verification:
>   - Your normal Anderson--Darling test does not agree with nortest::ad.test. Why is that so?
>

Commit [bc6d412](https://github.com/ComtekAdvancedStructures/cmstatr/commit/bc6d4126b3ecdcf7541db9cdf88aa9e18dfe8e1f). There is a modification, which depends on sample size and distribution, made to the test statistic. This modification is intended to account for the fact that the population parameters are estimated from the sample, rather than known. There are at least two sets of the modifications in the literature. `nortest` uses the modification formula from D'Agostino (1986), while `cmstatr` uses the modification formula from Stephens (1974). The unmodified test statistic is reported by both `nortest` and `cmstatr` and for all the examples that I've tried, these agree. However, the p-value is computed based on the modified test statistic. Both `nortest` and `cmstatr` also use different equations to calculate the p-value based on their respective modified test statistics, so the resulting p-values are different. I've added some text to the details section of the `anderson_darling_normal` function about this. The method used in `cmstatr` matches the method published in CMH-17-1G and that is currently used in the software typically used for statistical analysis of composite material data, and I think that matching the results of the current software as much as possible has value.

>
> - Examples. It would be nice to have examples for all exported functions; this should be easy;
>   just copy from your test_that folder. In particular, the following functions lack examples.
>   - levene_test

Commit [668ee9b](https://github.com/ComtekAdvancedStructures/cmstatr/commit/668ee9be38e1a2afb7ef24c82bf5256725905e7d)

>
>   - maximum_normed_residual
>

Commit [92ffa21](https://github.com/ComtekAdvancedStructures/cmstatr/commit/92ffa218c90976e9a706d6a0a4f750ad0a3a6376)

>
>   - nonpara_binomial_rank
>

Commit [608a849](https://github.com/ComtekAdvancedStructures/cmstatr/commit/608a8496a0f4e76f9411c8b38d4910dbcd3e6594). This commit also added or improved the examples for the functions:

- basis_normal
- basis_pooled_sd
- calc_cv_star
- cv
- hk_ext
- k_equiv
- k_factor_normal
- normalize_group_mean
- normalize_ply_thickness

# Installation

> Calling
> ```
> devtools::install_github("ComtekAdvancedStructures/cmstatr", build_vignettes = TRUE, build_opts = c("--no-resave-data", "--no-manual"))
> ```
>
> fails when rmarkdown dplyr, and tidyr aren't installed. I suggest adding
> install.packages("rmarkdown"), and the same for the rest, to the installation
> instructions or, (ii) add rmarkdown to Imports instead of Suggests.
> (iii) It is possible to add checks in the vignette code to find out if a
> package has been installed, then ask to the user to install it right there
> and then. Don't do it automatically though -- CRAN won't accept it!
> In my opinion, this third option is not worth the trouble.
>
> - Installation. Fix the rmarkdown, dplyr, tidyr issue.

Commit [7dc781b](https://github.com/ComtekAdvancedStructures/cmstatr/commit/7dc781bcdfec2295085c2fa9ed4ab0e5df18d83d) addressed this.

Installation was verified through fresh install of rocker/rstudio docker container as follows:

```
docker run -d -p 8787:8787 -e ROOT=TRUE rocker/rstudio
```

Install system dependencies inside docker container:

```
docker ps  # note container ID
docker exec -it <container ID> bash
sudo apt install zlib1g-dev libxml2-dev
```

Then, launch http://localhost:8787

I've confirmed that when `dplyr` and `tidyr` are not explicitly installed, the installation of `cmstatr` fails (as you reported). Executing the new installation instructions in README succeeds.

# Paper

> > Design Values are determined such that, with 95% confidence,
> > they are either the 99% or 90% one-sided lower confidence bound
> > of the material strength, depending on the type of structure.
>
> I was strongly confused by this passage, and I think it should be
> reworded or removed. An option is to use only the language of tolerance
> intervals/bounds, and provide a reference and introduction to their definition
> and interpretation. Tolerance intervals are too convoluted to properly define in
> this paper (or the documentation), but they should be referenced by their proper
> name so that people (that is, people like me! :p) can easily check if they are
> implemented correctly. Moreover, I suggest to use the terminology "content" and
> "confidence level" for p and conf. Since you only care about lower confidence
> levels, I suggest you remove references to confidence intervals and replace them
> with confidence levels instead; it is confusing that several functions take conf
> as an argument but fails to return a confidence interval.

Commit [c36c520](https://github.com/ComtekAdvancedStructures/cmstatr/commit/c36c5207c29a36c11e884451084b4f182a8ee6ea): The confusing sentence in the paper has been reworked.

>
> It is nice that you repeatedly state in the documentation that p = 0.9 and
> conf = 0.95 corresponds to a B-Basis and so on, you should also state what
> they are supposed to calculate when p != 0.9 and conf != 0.95. (That is, the
> lower limits of a one-sided tolerance interval with content p and confidence level 0.95.)
>

I've added a sentence about this in the documentation for the `basis...` functions. While `cmstatr` is capable of computing tolerance bounds for other contents or confidence levels, these are seldom needed in practice. I don't think that there is a need to discuss this explicitly in the paper due to the infrequent need for practitioners in this field to compute tolerance bounds with different content and/or confidence.

>
> A reference for the to me surprising calculations used in e.g. basis_normal and
> k_factor_normal is Statistical Tolerance Regions: Theory, Applications, and
> Computation (Krishnamoorthy, Mathew, 2009), Section 1.2. But you probably know
> better references than I do!
>

Commit [eac7522](https://github.com/ComtekAdvancedStructures/cmstatr/commit/eac7522e3fa5c6a53ddcb7e26421edcc4d5fcaf9): I've added a reference to the Krishnamoorthy book to the documentation for the `k_factor_normal` and `basis_normal` functions. The "surprising" equation also appears in CMH-17-1G, but having more references certainly can't hurt.

>
> - Documenation: Reference to tolerance bounds.
>

I've added a reference to the `basis...` documentation and to the paper itself.

>
> The rest of the software paper looks fine to me!
>

I did notice, as I was addressing your comments, that the object returned by
the `maximum_normed_residual` function doesn't include the value of `alpha`
supplied by the user. To be consistent with the other functions, I've added
this in, and also now report the value of `alpha` used in the `print`
and `glance` methods for that class. See commit [c46e5e6](https://github.com/ComtekAdvancedStructures/cmstatr/commit/c46e5e630fec091426d233a329244ff60d06ad2d)
