### 2020-06-29 #2265 @JonasMoss
Thanks, @kloppen! Again, let me say I think this is a very nice package. =)

I think everything looks fine now, @usethedata.
