### 2020-06-29 7604e6f
Fixed lintr errors (two lines longer than 80 char)
