### 2020-06-29 4e18fc9
Updated NEWS.md and bumped version number
