+++ pkgdown.yml
@@ -1,11 +1,8 @@
 pandoc: 2.7.3
 pkgdown: 1.5.1.9000
-pkgdown_sha: aeb1cfb080f566b32da6613b6e12c13eb743daff
-articles:
-  adktest: adktest.html
-  cmstatr_Graphing: cmstatr_Graphing.html
-  cmstatr_Tutorial: cmstatr_Tutorial.html
-last_built: 2020-06-01T13:36Z
+pkgdown_sha: caf7cc74e008e8d0d6e53eeda93cd3cad3545bf2
+articles: []
+last_built: 2020-06-29T14:29Z
 urls:
   reference: www.cmstatr.net/reference
   article: www.cmstatr.net/articles
