### 2020-06-29 #2265 @kloppen
The changes that I made to address JonasMoss' comments were in the `feature/joss-review` branch of the `cmstatr` repository. I've now merged those changes into `master`.

I'll re-generate the PDF:

I did bump the software version number to `0.6.0` because there were some minor changes the resulted from the review. I think that only editors can ask whedon to change the version number. So, @usethedata, can you please update the version number for `cmstatr` to 0.6.0?

I haven't been through the process of publishing in JOSS before, so please let me know what, if anything, I need to do as the author at this point.
