### 2020-07-04 ComtekAdvancedStructures/cmstatr#19 @usethedata
Corrections/Actions for paper publication in JOSS
## Description

This issue lists corrections (nits), suggestions, and next step actions for publication of this paper in JOSS

## Corrections/Suggestions/nits

- [x] For the Raju and Tomblin reference, I suggest including a URL to the work.  It appears https://www.niar.wichita.edu/agate/ would be reasonable.
- [x] For the Scholz reference, Anderson-Darling should be capitalized

## Next steps

- [x] Merge the changes from above into your main branch.  Those are such minor nits that I would put them into the current version tagged release, rather than changing the version number.
- [x] Create a DOI archive of the software (e.g. on Figshare or Zenodo)
- [x] Post the DOI over on the [JOSS review issue](https://github.com/openjournals/joss-reviews/issues/2265).  Post the version number if you changed that.
