### 2020-07-04 #2265 @usethedata
@kloppen I've created an issue in your repository with a couple of nits to consider and the next steps in the process: https://github.com/ComtekAdvancedStructures/cmstatr/issues/19
