### 2020-07-04 468c0e9
Removed macOS/devel from CI due to failing upstream package
