+++ .github/workflows/R-CMD-check.yaml
@@ -25,7 +25,7 @@ jobs:
       matrix:
         config:
         - { os: windows-latest, r: 'release',   vdiffr: true  }
-        - { os: macOS-latest,   r: 'devel', vdiffr: false }
+        # - { os: macOS-latest,   r: 'devel', vdiffr: false }
         - { os: macOS-latest,   r: 'release',   vdiffr: false }
         - { os: ubuntu-16.04,   r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", vdiffr: true  }
         - { os: ubuntu-16.04,   r: 'oldrel', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest", vdiffr: false }
