### 2020-07-04 ComtekAdvancedStructures/cmstatr#19 @kloppen
I've addressed the corrections/nits above and completed the "next steps."
