### 2020-07-04 #2265 @kloppen
@usethedata :

The Zenodo DOI is: 10.5281/zenodo.3930475

I've fixed the couple little nits that you raised on ComtekAdvancedStructures/cmstatr#19. I did not change software version number (so, it's still 0.6.0).

Please let me know if you need me to do anything else.
