### 2020-07-05 #2265 @danielskatz
@kloppen - I've suggested some small changes to the paper in https://github.com/ComtekAdvancedStructures/cmstatr/pull/20 - Once these are made/merged (or you let me know what you disagree with), we can finish the publishing of this work.
