### 2020-07-05 923157e
Update paper.Rmd to match commit e618e9d
