+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.7.3
 pkgdown: 1.5.1.9000
 pkgdown_sha: caf7cc74e008e8d0d6e53eeda93cd3cad3545bf2
 articles: []
-last_built: 2020-07-05T20:49Z
+last_built: 2020-07-05T20:54Z
 urls:
   reference: www.cmstatr.net/reference
   article: www.cmstatr.net/articles
