### 2020-07-05 #2265 @kloppen
Thanks for your corrections, @danielskatz. I've merged your PR (and then made a subsequent commit to roll those changes into the R-Notebook that generates the paper.md file; that subsequent commit is [923157e](https://github.com/ComtekAdvancedStructures/cmstatr/commit/923157e278a6a97c81a82221d4eaf413e82b22b3)).

Please let me know if there is anything else I need to do as the author.
